+++ lib/fizzy/formatter.rb
@@ -7,13 +7,13 @@ module Fizzy
 
       all_rows = [headers] + rows
       widths = headers.each_index.map do |i|
-        all_rows.map { |r| r[i].to_s.length }.max
+        all_rows.map { |r| display_width(r[i].to_s) }.max
       end
 
-      io.puts headers.each_with_index.map { |h, i| h.to_s.ljust(widths[i]) }.join("  ")
+      io.puts headers.each_with_index.map { |h, i| display_ljust(h.to_s, widths[i]) }.join("  ")
       io.puts widths.map { |w| "-" * w }.join("  ")
       rows.each do |row|
-        io.puts row.each_with_index.map { |c, i| c.to_s.ljust(widths[i]) }.join("  ")
+        io.puts row.each_with_index.map { |c, i| display_ljust(c.to_s, widths[i]) }.join("  ")
       end
     end
 
@@ -33,5 +33,27 @@ module Fizzy
 
       str.length > max ? "#{str[0...(max - 1)]}…" : str
     end
+
+    def self.display_width(str)
+      str = str.to_s
+      str.each_char.sum { |c| wide_char?(c) ? 2 : 1 }
+    end
+
+    def self.display_ljust(str, width)
+      str = str.to_s
+      padding = width - display_width(str)
+      padding.positive? ? "#{str}#{" " * padding}" : str
+    end
+
+    WIDE_RANGES = [
+      0x1100..0x115F, 0x2E80..0xA4CF, 0xAC00..0xD7AF, 0xF900..0xFAFF,
+      0xFE10..0xFE6F, 0xFF00..0xFF60, 0x1F000..0x1FFFF, 0x20000..0x2FA1F
+    ].freeze
+
+    def self.wide_char?(char)
+      WIDE_RANGES.any? { |r| r.cover?(char.ord) }
+    end
+
+    private_class_method :display_width, :display_ljust, :wide_char?
   end
 end